AngularJS
I added all of the language files for AngularJS. So, most things
translate from English to Italian. The problem is that the PHP form
doesn’t submit if I have the angular tags…. Need to fix that. Then, I
need to figure out how to translate the tooltips and some other
details. I also need to fix the width of the page so that it doesn’t
show a gap. Not the final project yet.