Back from NG route
On the way back from NG Route. Still having problems with the select
room function and dates function on the index page. Need to do a lot of
work on the contact page as well for the contact form. A lot of things
aren’t compatible with Angular (jQuery in particular). So, I need to
find a few work arounds.